requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
aiosqlite==0.17.0
annotated-types==0.7.0
async-timeout==4.0.3
asyncio==3.4.3
asyncpg==0.29.0
atlastk==0.13.3
attrs==24.2.0
black==24.8.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
frozenlist==1.4.1
idna==3.10
iso8601==1.1.0
multidict==6.1.0
mypy-extensions==1.0.0
packaging==24.1
pathspec==0.12.1
platformdirs==4.3.3
pydantic==2.9.1
pydantic_core==2.23.3
pypika-tortoise==0.1.6
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
schedule==1.2.2
tomli==2.0.1
tortoise==0.1.1
tortoise-orm==0.21.6
typing_extensions==4.12.2
urllib3==2.2.3
yarl==1.11.1
